Fix dependencies issues
