+++ core/application.go
@@ -15,7 +15,6 @@
 package core
 
 import (
-	"application/core/interfaces"
 	"application/core/model"
 
 	"github.com/rokwire/core-auth-library-go/v3/authutils"
@@ -41,21 +40,21 @@ type Application struct {
 	version string
 	build   string
 
-	Default interfaces.Default // expose to the drivers adapters
-	Client  interfaces.Client  // expose to the drivers adapters
-	Admin   interfaces.Admin   // expose to the drivers adapters
-	BBs     interfaces.BBs     // expose to the drivers adapters
-	TPS     interfaces.TPS     // expose to the drivers adapters
-	System  interfaces.System  // expose to the drivers adapters
+	Default Default // expose to the drivers adapters
+	Client  Client  // expose to the drivers adapters
+	Admin   Admin   // expose to the drivers adapters
+	BBs     BBs     // expose to the drivers adapters
+	TPS     TPS     // expose to the drivers adapters
+	System  System  // expose to the drivers adapters
 	shared  Shared
 
-	AppointmentAdapters map[string]interfaces.Appointments //expose to the different vendor specific appointment adapters
+	AppointmentAdapters map[string]Appointments //expose to the different vendor specific appointment adapters
 
 	logger *logs.Logger
 
-	storage interfaces.Storage
+	storage Storage
 
-	eventsBBAdapter interfaces.EventsBBAdapter
+	eventsBBAdapter EventsBBAdapter
 
 	//events logic
 	eventsLogic eventsLogic
@@ -91,9 +90,9 @@ func (a *Application) GetEnvConfigs() (*model.EnvConfigData, error) {
 
 // NewApplication creates new Application
 func NewApplication(version string, build string,
-	storage interfaces.Storage,
-	eventsBBAdapter interfaces.EventsBBAdapter,
-	appntAdapters map[string]interfaces.Appointments,
+	storage Storage,
+	eventsBBAdapter EventsBBAdapter,
+	appntAdapters map[string]Appointments,
 	logger *logs.Logger) *Application {
 	application := Application{version: version, build: build, storage: storage, eventsBBAdapter: eventsBBAdapter, logger: logger, AppointmentAdapters: appntAdapters}
